+++ libs/source-panel/src/components/ConversationSourcesPanel/ConversationSourcesPanel.tsx
@@ -1,4 +1,3 @@
-import type { DisplayAttachment } from '@epam/ai-dial-chat-shared';
 import {
   PanelEmpty,
   PanelNoResults,
@@ -9,79 +8,14 @@ import {
 import { DIAL_ICON_SIZE, DialGhostIconButton } from '@epam/ai-dial-ui-kit';
 import { IconDownload } from '@tabler/icons-react';
 import { memo, useLayoutEffect, useMemo, useState, type FC } from 'react';
-import type { QuotationSource } from '../../models/quotation-source';
-import FilesSection from './sections/FilesSection/FilesSection';
-import SourcesSection from './sections/SourcesSection/SourcesSection';
-
-/** User-visible strings passed by the consuming app (via i18n or plain strings). */
-export interface ConversationSourcesPanelLabels {
-  /** Accessible name for the sidebar panel element. */
-  ariaLabel: string;
-  /** Label for the close button. */
-  closeLabel: string;
-  /** Placeholder text shown inside the search input. */
-  searchPlaceholder: string;
-  /** Accessible label for the search input clear button. */
-  searchClearLabel: string;
-  /** Text shown when the panel has no files or sources. */
-  emptyLabel: string;
-  /** Text shown when a search query matches nothing. */
-  noResultsLabel: string;
-  /** Accessible label for the download-all icon button. */
-  downloadAllLabel: string;
-  /** Heading text for the uploaded-files section. */
-  uploadedSectionTitle: string;
-  /** Heading text for the generated-files section. */
-  generatedSectionTitle: string;
-  /** Heading text for the sources section. */
-  sourcesSectionTitle: string;
-  /** Accessible label for each source's copy-URL button. */
-  copySourceLabel: string;
-  /** Label passed to each attachment card's action button. */
-  attachmentClickLabel: string;
-}
-
-/** Optional typography and color class overrides for inner elements. */
-export interface ConversationSourcesPanelStyles {
-  /** CSS class applied to each section heading (`<h2>`). Defaults to `'dial-body-semi-text'`. */
-  sectionTitleClassName?: string;
-  /** CSS class applied to each source link. Defaults to `'dial-small-text !text-accent-primary'`. */
-  sourceLinkClassName?: string;
-  /** CSS class applied to each source quote paragraph. Defaults to `'dial-tiny-text text-secondary'`. */
-  sourceQuoteClassName?: string;
-}
-
-/** Props for the ConversationSourcesPanel component. */
-export interface ConversationSourcesPanelProps {
-  /** Whether the panel is currently visible. */
-  isOpen: boolean;
-  /** Called when the user requests to close the panel. */
-  onClose: () => void;
-  /** Attachments uploaded by the user in this conversation. */
-  uploaded: DisplayAttachment[];
-  /** Attachments generated by the assistant in this conversation. */
-  generated: DisplayAttachment[];
-  /** Quotation sources cited by the assistant. */
-  sources: QuotationSource[];
-  /** Called when the user activates an attachment card. */
-  onAttachmentClick?: (attachment: DisplayAttachment) => void;
-  /** Called when the user clicks a source link. When omitted, the link navigates normally. */
-  onSourceClick?: (source: QuotationSource) => void;
-  /** Whether the viewport is in mobile breakpoint (disables resizing). */
-  isMobile: boolean;
-  /** Initial panel width in pixels. */
-  defaultWidth?: number;
-  /** Minimum panel width in pixels. */
-  minWidth?: number;
-  /** Maximum panel width in pixels. */
-  maxWidth?: number;
-  /** Called with the new width after the user finishes a resize drag. */
-  onResizeStop?: (width: number) => void;
-  /** User-visible strings (titles, placeholders, aria labels). */
-  labels: ConversationSourcesPanelLabels;
-  /** Optional typography and color class overrides for inner elements. */
-  styles?: ConversationSourcesPanelStyles;
-}
+import type { ConversationSourcesPanelProps } from '../../models/conversation-sources-panel-props';
+import FilesSection from '../FilesSection/FilesSection';
+import SourcesSection from '../SourcesSection/SourcesSection';
+export type {
+  ConversationSourcesPanelLabels,
+  ConversationSourcesPanelProps,
+  ConversationSourcesPanelStyles,
+} from '../../models/conversation-sources-panel-props';
 
 const includesIgnoreCase = (text: string, query: string): boolean =>
   text.toLowerCase().includes(query.toLowerCase());
