+++ libs/source-panel/src/components/FilesSection/FilesSection.tsx
@@ -3,12 +3,17 @@ import { AttachmentType, mergeClasses } from '@epam/ai-dial-chat-shared';
 import { AttachmentCard } from '@epam/ai-dial-conversation-input';
 import { memo, type FC } from 'react';
 
+/** Props for the FilesSection component. */
 interface FilesSectionProps {
+  /** List of attachments to display. Renders nothing when empty. */
   attachments: DisplayAttachment[];
+  /** Heading text rendered above the attachment grid. */
   title: string;
   /** CSS class applied to the section heading. Defaults to `'dial-body-semi-text'`. */
   titleClassName?: string;
+  /** Called when the user clicks an attachment card. */
   onAttachmentClick?: (attachment: DisplayAttachment) => void;
+  /** Accessible label for the attachment click action, forwarded to `AttachmentCard`. */
   attachmentClickLabel?: string;
 }
