+++ libs/source-panel/src/components/FilesSection/tests/FilesSection.spec.tsx
